9bootpbs: make even smaller
9bootpbs is loaded at 0x10000 and then uses a little extra memory
and needs to end before 512k so that the memory tables can be
copied there. It was too large, producing a "weird low-memory map"
error message and then amd64 kernels that panicked at startup
because they had no memory.

I already deleted most of 9bootpbs earlier to make it fit in 640k.

The old size plus 64k was under 512k, but empirically was still too big.
This one works.