sys/src/9k: fix noted(NCONT) losing registers
If a process is interrupted by a trap and decides on the way
out of the trap to handle a pending note, then when the
note handler finishes and calls noted(NCONT), the kernel
must return back to the original process context using IRET
in order to restore all the registers to their exact state at the
time of the trap. (This is the only case where a system call
does not return to the code immediately after the SYSCALL
instruction.)

The existing system call handler code assumed that it could
always exit a syscall using SYSRET; that is, it assumed that
it would always return from a syscall back to immediately
after the SYSCALL instruction.

As a result, the noted(NCONT) case just described used
SYSRET and corrupted some of the registers on its way back.

It is difficult to believe the system worked as well as it did.
I guess most programs do not noted(NCONT).